requirements.txt
flask
numpy
pillow
torch
torchvision
opencv-python-headless
gunicorn
